requirements.txt
streamlit>=1.32.0
google-generativeai>=0.8.0
google-genai>=1.10.0
pandas>=2.0.0
requests>=2.31.0
openpyxl>=3.1.0